requirement.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
certifi==2024.8.30
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jiter==0.8.0
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26
openai==1.57.0
packaging
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==5.9.3
psycopg==3.2.3
psycopg-binary==3.2.3
pure_eval==0.2.3
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
streamlit==1.51
python-dotenv


# Database and storage
psycopg[binary]
psycopg-pool
psutil

# Document processing
pandoc
pypandoc
markdown-it-py
langchain-text-splitters
beautifulsoup4

# Utilities
python-magic
aiohttp
requests
PyYAML
tenacity
tqdm
rich

llama-index 
llama-index-core 
llama-index-llms-openai 
llama-index-program-openai 
llama-index-question-gen-openai
pydantic_settings
